install: new no dereference option
install like cp support -P or --no-dereference. The installation doesn't follow the symlinks and creates a new one.
The file "copy.c" already contains the code to do it. Only install.c has to set some options to enable the feature.
After that copy_internal creates a symlink, install.c hasn't to change access mode on this file, because the mode is applied on the target of the symlink. This target may not exist at this moment.