install: install directory recursively.
install accepts the -r --recursive option like cp, to install the contains of a SOURCE directory to a DEST directory.
copy contains the main code of the recursive copy. A problem exists only with directory when the set_mode option is activated. The mode option is fixed with install for all SOURCE files (directories). But for directory the mode must add at least the ---x------ access right to the directories.