scripts: new TRIPLET variable
TRIPLET is the target configuration name for a lot of software
It's defined by <ARCH>-<KERNEL>-<LIBC><EXTRA>
Example for arm target with linux kernel and glibc the TRIPLET is
arm-linux-gnueabi
The triplet is often similar to the CROSS_COMPILE, but CROSS_COMPILE is
the TRIPLET used during the toolchain build. Here TRIPLET is for the
rootfs. TRIPLET is used for host argument of a lot of packages, but CC
must be defined with <CROSS_COMPILE>gcc otherwise the packages will try
to use <TRIPLET>gcc which can not exist.