scripts: libgcc is required for any GCC build
GCC build binary with some codes available outside the binary. This code is stored inside libgcc. This library is shared or static but in each case the binary must be linked to libgcc.
at the run time, libgcc_h.so is required when you start an application build with GCC.