scripts: gcc wrapper to use specs file
The specs file must be pass only once to the gcc command line.
But libtool some time use only CFLAGS, or CPPFLAGS or LDFLAGS or a combinaison of the threes. Then it's impossible to insert inside only one and not inside the threes.
specs-wrapper-gcc in a script that calls gcc with specs file as argument.
specs-wrapper-gcc must be installed inside out/host/toolchain/bin